Add host network and volumes to docker-compose
The "network_mode" and "volumes" properties are now added to the docker-compose.yml file for the container "smart-home". This change will allow the container to share network with the host to be able to listen to multicast UDP messages and persist app configurations and logs for easier debugging and configuration change without rebuilding the image.