Update Makefile to separate build and deploy stages
Jobs 'build' and 'deploy' have been separated in the Makefile for clarity and improved management. The 'build' job now handles the local build, creating a smart-home image tagged with 'alpha-1'. The 'deploy' job then handles the deployment to 'raspberrypi'. Also created a dedicated directory for configs and cache on the raspberry pi machine as part of the 'deploy' job, to improve organization. These changes provide a clearer division of responsibilities in the Makefile.