Refactor Telegram Bot and add new extensions
Refactored Telegram Bot to improve encapsulation and to allow it in other extensions. A new 'sun-telegram-notification' extension was added to send notifications at sunrise and sunset. Dockerfile was also updated - '.env' file is no longer copied, environment variables should be passed in runtime.
The DeviceClassRegistry's register function now throws an Error instead of returning false. Sun device class was updated - renamed to SunDevice and is now registered in DeviceClassRegistry via the 'sun' extension. Fixes were applied to comply with the new usage of DeviceClassRegistry and TelegramBot.