Add extension ordering based on dependencies
Introduced a new method, `dependsOn()`, as part of Extension abstract class to clearly define dependencies for each specific extension. This method returns an array of strings that represent the dependencies of the extension. Altered `extensions-loader` to sort extensions based on the above dependencies, ensuring proper sequence of extension execution. This will prevent potential issues  in dependent extensions.