Change Cube's rotation effects on Bulb device
Enhanced the Cube's interaction with the Bulb device on its rotations. The left and right rotations now adjust (decrease and increase respectively) the brightness and color temperature of the Bulb. If the cube is flipped 90 degree, the mode switches between brightness and color temperature. A logger error is added when no rotation angle is detected. Device is fetched in the beginning to prevent multiple fetches and log an error if not found.